+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * Latest development version.
 * Added the `TestStationarity()`, `TestStability()`, and `TestPhi()` functions.
-* Added the `SimBeta()` and `SimPhi()` functions.
+* Added the `SimBetaN()` and `SimPhiN()` functions.
 * The `SimSSMLinSDEIVary()` and `SimSSMOUIVary()` functions can now accept a matrix of zeros for the argument `sigma_l`.
 * The `SimSSMLinSDEIVary()` function can now accept a vector of zeros for the argument `iota`.
 * The `SimSSMOUIVary()` function can now accept a vector of zeros for the argument `mu`.